Conditional Distribution Function of a Bivariate Copula

Evaluate the conditional distribution function (h-function) of a given parametric bivariate copula.

BiCopHfunc(u1, u2, family, par, par2 = 0, obj = NULL, check.pars = TRUE)

BiCopHfunc1(u1, u2, family, par, par2 = 0, obj = NULL, check.pars = TRUE)

BiCopHfunc2(u1, u2, family, par, par2 = 0, obj = NULL, check.pars = TRUE)

Arguments

u1, u2

numeric vectors of equal length with values in \([0,1]\).

family

integer; single number or vector of size length(u1); defines the bivariate copula family:
0 = independence copula
1 = Gaussian copula
2 = Student t copula (t-copula)
3 = Clayton copula
4 = Gumbel copula
5 = Frank copula
6 = Joe copula
7 = BB1 copula
8 = BB6 copula
9 = BB7 copula
10 = BB8 copula
13 = rotated Clayton copula (180 degrees; survival Clayton'') \cr `14` = rotated Gumbel copula (180 degrees; survival Gumbel'')
16 = rotated Joe copula (180 degrees; survival Joe'') \cr `17` = rotated BB1 copula (180 degrees; survival BB1'')
18 = rotated BB6 copula (180 degrees; survival BB6'')\cr `19` = rotated BB7 copula (180 degrees; survival BB7'')
20 = rotated BB8 copula (180 degrees; ``survival BB8'')
23 = rotated Clayton copula (90 degrees)
`24` = rotated Gumbel copula (90 degrees)
`26` = rotated Joe copula (90 degrees)
`27` = rotated BB1 copula (90 degrees)
`28` = rotated BB6 copula (90 degrees)
`29` = rotated BB7 copula (90 degrees)
`30` = rotated BB8 copula (90 degrees)
`33` = rotated Clayton copula (270 degrees)
`34` = rotated Gumbel copula (270 degrees)
`36` = rotated Joe copula (270 degrees)
`37` = rotated BB1 copula (270 degrees)
`38` = rotated BB6 copula (270 degrees)
`39` = rotated BB7 copula (270 degrees)
`40` = rotated BB8 copula (270 degrees)
`104` = Tawn type 1 copula
`114` = rotated Tawn type 1 copula (180 degrees)
`124` = rotated Tawn type 1 copula (90 degrees)
`134` = rotated Tawn type 1 copula (270 degrees)
`204` = Tawn type 2 copula
`214` = rotated Tawn type 2 copula (180 degrees)
`224` = rotated Tawn type 2 copula (90 degrees)
`234` = rotated Tawn type 2 copula (270 degrees)

par

numeric; single number or vector of size length(u1); copula parameter.

par2

numeric; single number or vector of size length(u1); second parameter for bivariate copulas with two parameters (t, BB1, BB6, BB7, BB8, Tawn type 1 and type 2; default: par2 = 0). par2 should be an positive integer for the Students's t copula family = 2.

obj

BiCop object containing the family and parameter specification.

check.pars

logical; default is TRUE; if FALSE, checks for family/parameter-consistency are omitted (should only be used with care).

Value

BiCopHfunc returns a list with

hfunc1

Numeric vector of the conditional distribution function (h-function) of the copula family with parameter(s) par, par2 evaluated at u2 given u1, i.e., \(h_1(u_2|u_1;\boldsymbol{\theta})\).

hfunc2

Numeric vector of the conditional distribution function (h-function) of the copula family with parameter(s) par, par2 evaluated at u1 given u2, i.e., \(h_2(u_1|u_2;\boldsymbol{\theta})\).

BiCopHfunc1 is a faster version that only calculates hfunc1; BiCopHfunc2 only calculates hfunc2.

Details

The h-function is defined as the conditional distribution function of a bivariate copula, i.e., $$h_1(u_2|u_1;\boldsymbol{\theta}) := P(U_2 \le u_2 | U_1 = u_1) = \frac{\partial C(u_1, u_2; \boldsymbol{\theta})}{\partial u_1}, $$ $$h_2(u_1|u_2;\boldsymbol{\theta}) := P(U_1 \le u_1 | U_2 = u_2) = \frac{\partial C(u_1, u_2; \boldsymbol{\theta})}{\partial u_2}, $$ where \((U_1, U_2) \sim C\), and \(C\) is a bivariate copula distribution function with parameter(s) \(\boldsymbol{\theta}\). For more details see Aas et al. (2009).

If the family and parameter specification is stored in a BiCop() object obj, the alternative versions

BiCopHfunc(u1, u2, obj)
BiCopHfunc1(u1, u2, obj)
BiCopHfunc2(u1, u2, obj)

can be used.

References

Aas, K., C. Czado, A. Frigessi, and H. Bakken (2009). Pair-copula constructions of multiple dependence. Insurance: Mathematics and Economics 44 (2), 182-198.

See also

Author

Ulf Schepsmeier

Examples

data(daxreturns)

# h-functions of the Gaussian copula
cop <- BiCop(family = 1, par = 0.5)
h <- BiCopHfunc(daxreturns[, 2], daxreturns[, 1], cop)
h
#> $hfunc1
#>    [1] 6.220108e-01 3.218472e-01 7.129364e-01 4.437906e-01 5.492441e-01
#>    [6] 2.659561e-01 3.172890e-01 5.424892e-01 4.214900e-01 5.596056e-01
#>   [11] 1.808212e-01 2.110578e-02 3.268378e-01 4.186970e-01 4.539761e-01
#>   [16] 8.819727e-01 7.710018e-02 2.066970e-01 4.570489e-01 7.192216e-01
#>   [21] 3.625526e-01 7.493323e-01 4.091462e-01 3.606287e-01 9.520267e-01
#>   [26] 3.580507e-01 1.260451e-01 3.992607e-01 8.158813e-01 5.235301e-01
#>   [31] 3.404628e-01 9.973454e-02 3.719666e-01 1.666953e-02 8.525902e-01
#>   [36] 5.271848e-01 3.355021e-01 9.095176e-01 8.125316e-01 5.717504e-01
#>   [41] 6.645106e-01 6.481941e-01 3.107388e-01 7.543199e-01 3.066058e-01
#>   [46] 8.879242e-01 1.261062e-01 5.370029e-01 9.911515e-01 8.153369e-01
#>   [51] 7.832804e-01 1.739493e-01 1.891788e-01 7.224191e-01 5.028833e-01
#>   [56] 4.522016e-01 2.189110e-01 9.199584e-01 7.869646e-01 4.762817e-01
#>   [61] 8.638222e-02 4.252703e-01 1.049839e-01 3.728218e-01 2.958434e-01
#>   [66] 6.637995e-01 3.207335e-01 4.260785e-01 1.753042e-01 9.926601e-01
#>   [71] 7.040399e-01 5.598386e-02 7.551352e-02 2.291185e-01 1.864402e-01
#>   [76] 8.521182e-01 6.415018e-01 6.317190e-01 3.704021e-01 7.012655e-01
#>   [81] 2.146725e-01 9.657149e-01 6.505207e-01 6.852600e-01 6.963995e-01
#>   [86] 2.094733e-01 5.398654e-01 3.378816e-01 2.054165e-02 1.517736e-01
#>   [91] 7.462270e-01 6.324766e-01 6.053802e-01 5.225672e-01 6.995091e-01
#>   [96] 7.088544e-01 1.418718e-01 3.575292e-01 2.761821e-01 5.208735e-01
#>  [101] 8.470385e-01 3.948429e-01 7.485560e-01 8.005309e-01 4.848418e-01
#>  [106] 2.013726e-01 2.712788e-01 8.256906e-02 8.525972e-01 5.032331e-01
#>  [111] 6.371786e-01 4.639744e-01 4.729835e-01 3.866720e-01 3.431906e-01
#>  [116] 6.087868e-01 6.601683e-01 1.874756e-01 8.335392e-01 8.003653e-01
#>  [121] 3.665657e-01 1.947257e-01 5.881205e-02 4.375514e-01 9.411151e-01
#>  [126] 1.332270e-01 4.419045e-01 5.477902e-01 6.204754e-01 9.974483e-01
#>  [131] 8.600678e-02 9.031677e-01 6.408202e-01 2.446886e-01 5.065390e-01
#>  [136] 2.607399e-01 6.340757e-01 5.256622e-01 5.211369e-01 7.913579e-01
#>  [141] 4.546972e-01 7.202678e-01 7.304427e-01 6.881829e-01 8.335851e-01
#>  [146] 7.158278e-01 1.998259e-01 8.179546e-01 5.276375e-01 1.513553e-01
#>  [151] 3.252458e-01 4.025327e-01 5.493411e-01 9.425643e-01 9.942326e-01
#>  [156] 3.825682e-01 4.242957e-01 6.343864e-01 9.263601e-01 9.176409e-02
#>  [161] 3.030937e-01 6.559799e-01 6.485043e-01 1.223301e-01 5.428482e-01
#>  [166] 9.212406e-02 1.610594e-01 3.656983e-01 5.104640e-01 8.445684e-01
#>  [171] 7.804541e-02 1.599773e-01 8.064477e-01 7.440593e-01 6.936814e-01
#>  [176] 7.822910e-01 3.743332e-01 4.744443e-04 4.002331e-01 7.381080e-01
#>  [181] 7.904385e-01 5.868182e-01 2.399785e-01 5.398514e-01 3.724882e-01
#>  [186] 3.228811e-01 6.624500e-01 8.126232e-01 5.463096e-01 9.512261e-01
#>  [191] 4.814837e-01 5.301889e-01 2.147525e-01 6.535847e-01 2.559981e-01
#>  [196] 3.579175e-01 8.283286e-01 8.870790e-01 8.341937e-01 5.903904e-01
#>  [201] 4.807299e-01 8.526493e-01 8.255190e-01 2.282090e-01 1.320756e-01
#>  [206] 7.775126e-01 6.283700e-01 5.481352e-01 9.123231e-01 6.748970e-01
#>  [211] 2.689369e-01 3.319067e-01 8.463019e-01 7.424585e-01 8.283626e-01
#>  [216] 6.119915e-01 4.210478e-01 4.769737e-01 4.135918e-01 2.903747e-01
#>  [221] 9.279669e-01 1.832669e-02 3.997622e-01 4.360610e-01 6.157701e-01
#>  [226] 6.775670e-01 9.367123e-01 2.842308e-01 4.739750e-01 2.813141e-01
#>  [231] 4.553210e-01 7.070964e-01 7.987352e-01 7.299530e-01 3.025411e-01
#>  [236] 9.069150e-01 9.608043e-01 8.180575e-01 5.125825e-01 6.650703e-02
#>  [241] 1.718663e-01 4.635191e-01 8.976720e-01 2.922489e-01 1.300253e-01
#>  [246] 5.460797e-01 6.306756e-01 8.098946e-01 2.642022e-01 7.639864e-01
#>  [251] 2.931063e-01 5.775031e-01 7.860199e-01 3.518381e-01 2.794586e-01
#>  [256] 2.858959e-01 8.276341e-01 9.808762e-01 9.996882e-01 4.336880e-01
#>  [261] 9.485243e-01 4.278432e-01 1.034282e-01 4.728999e-01 1.727999e-01
#>  [266] 7.037375e-02 4.156002e-01 7.386727e-02 2.945748e-02 9.365961e-01
#>  [271] 3.764979e-01 4.965403e-01 1.908080e-01 8.787847e-01 9.697081e-01
#>  [276] 5.137319e-01 6.952720e-01 1.381933e-01 6.195187e-02 1.863377e-01
#>  [281] 6.696049e-01 7.530222e-01 6.653448e-01 6.788227e-01 6.857185e-01
#>  [286] 5.410656e-01 6.399272e-01 4.589897e-01 4.850110e-01 9.146477e-01
#>  [291] 8.445714e-01 3.214657e-01 1.510498e-01 8.928313e-01 7.094392e-01
#>  [296] 5.331179e-01 7.514957e-01 2.164223e-01 8.920378e-01 8.928208e-02
#>  [301] 3.631122e-01 8.404453e-01 3.431239e-02 1.389114e-02 6.617135e-01
#>  [306] 8.001245e-01 9.449664e-01 7.558392e-01 7.800364e-01 2.303142e-02
#>  [311] 7.833009e-01 1.940282e-01 9.266794e-01 2.013589e-01 3.289097e-01
#>  [316] 3.948048e-02 7.211418e-01 6.137407e-01 9.435936e-01 7.313730e-01
#>  [321] 2.006082e-01 6.831126e-01 5.633407e-01 6.980575e-01 1.107408e-01
#>  [326] 1.195294e-01 3.272125e-01 9.111995e-02 3.833568e-01 2.031279e-01
#>  [331] 7.020209e-01 3.549751e-01 4.287367e-01 1.339121e-02 4.483916e-02
#>  [336] 6.283449e-01 5.391445e-01 9.575830e-01 1.552065e-01 6.634129e-01
#>  [341] 1.577134e-01 7.916594e-02 9.987645e-01 9.143786e-01 2.176331e-01
#>  [346] 3.867223e-01 2.338364e-01 1.107013e-01 3.703267e-01 4.761017e-01
#>  [351] 3.593191e-02 4.432781e-01 3.788068e-01 1.830582e-01 7.605679e-01
#>  [356] 1.295962e-01 4.877590e-01 6.653828e-01 3.868824e-01 1.713247e-01
#>  [361] 8.742815e-01 4.473795e-01 4.692181e-01 5.478841e-01 7.072362e-02
#>  [366] 4.864522e-01 1.644903e-01 8.619222e-01 4.230570e-01 1.670347e-01
#>  [371] 5.481141e-01 9.955785e-01 5.022949e-01 4.363372e-01 7.205906e-01
#>  [376] 2.112372e-01 5.188457e-01 4.617394e-01 6.250282e-01 6.960037e-01
#>  [381] 4.014993e-01 4.466227e-01 5.207871e-01 2.781201e-01 6.753447e-01
#>  [386] 4.572756e-01 3.490790e-01 2.479615e-01 3.402127e-01 1.582112e-01
#>  [391] 1.992836e-01 4.415896e-01 4.133930e-01 9.441801e-01 5.975012e-01
#>  [396] 4.854102e-01 7.633120e-01 2.977414e-01 4.622998e-01 7.335177e-01
#>  [401] 8.054686e-01 2.234338e-01 2.553034e-01 9.160890e-01 6.232499e-01
#>  [406] 9.899361e-01 4.083644e-01 6.087581e-01 5.210107e-01 6.377686e-01
#>  [411] 7.958881e-01 7.464691e-01 4.952654e-01 4.611044e-01 5.340745e-01
#>  [416] 3.303502e-01 4.747677e-01 1.620419e-01 4.709747e-01 6.509053e-01
#>  [421] 3.154967e-01 5.631926e-01 5.118057e-01 3.410325e-01 8.158386e-01
#>  [426] 5.948009e-01 7.715316e-01 3.471733e-01 4.128610e-01 2.112068e-01
#>  [431] 7.411379e-01 8.996246e-01 8.957563e-01 4.693744e-01 6.585784e-01
#>  [436] 6.894635e-01 3.937343e-01 5.424866e-01 4.330502e-01 3.417384e-01
#>  [441] 7.161468e-01 9.156758e-01 4.081541e-01 3.362508e-01 3.528603e-01
#>  [446] 1.690117e-01 5.061618e-01 9.243871e-01 8.758714e-01 4.997471e-01
#>  [451] 3.092422e-01 9.628350e-01 6.415034e-01 8.646630e-01 4.526095e-01
#>  [456] 4.557355e-02 4.631753e-01 2.065046e-01 7.361521e-01 5.369416e-01
#>  [461] 4.882243e-01 2.633679e-01 4.050061e-01 1.789975e-01 5.493113e-01
#>  [466] 8.915581e-01 3.067863e-01 3.350031e-01 9.830739e-01 8.313372e-01
#>  [471] 6.844965e-01 7.854408e-01 2.550006e-02 7.379854e-01 1.527453e-01
#>  [476] 4.423608e-01 2.386506e-01 2.840459e-01 9.602914e-01 3.141736e-01
#>  [481] 7.620242e-01 8.501982e-02 4.570668e-02 5.319074e-02 7.071177e-01
#>  [486] 7.765477e-01 3.632199e-01 2.993536e-01 5.237889e-01 7.598125e-01
#>  [491] 5.102885e-01 6.536787e-01 1.700186e-01 5.224210e-01 4.596340e-01
#>  [496] 7.963172e-01 9.019029e-01 4.303389e-01 9.695532e-01 3.504947e-01
#>  [501] 4.615124e-01 2.473633e-01 3.610598e-01 4.826241e-01 3.911701e-01
#>  [506] 2.916431e-01 6.819670e-01 8.374387e-01 9.779651e-01 1.158332e-01
#>  [511] 6.159596e-01 4.163163e-02 1.539914e-01 8.937990e-01 4.234580e-01
#>  [516] 4.936732e-01 2.465533e-01 1.490410e-01 1.110573e-01 5.564417e-01
#>  [521] 2.345601e-01 5.333389e-01 8.535397e-01 4.120595e-01 5.531162e-02
#>  [526] 8.821365e-01 7.913602e-01 1.996799e-01 2.788708e-01 7.574727e-01
#>  [531] 6.128369e-01 2.089941e-01 7.225543e-01 4.815936e-01 3.745438e-01
#>  [536] 9.423399e-02 6.440581e-01 7.649582e-01 5.066363e-01 4.380851e-01
#>  [541] 3.803321e-01 8.994935e-01 9.246082e-01 7.960675e-01 9.562377e-01
#>  [546] 7.165446e-02 6.881931e-01 7.994153e-02 2.335582e-01 3.574490e-01
#>  [551] 5.061341e-01 2.461889e-01 6.849919e-01 4.076003e-01 2.670269e-01
#>  [556] 3.445565e-01 3.868300e-02 4.786485e-01 4.449749e-01 7.072325e-01
#>  [561] 4.052631e-01 3.537385e-01 9.518808e-01 5.120865e-01 5.285303e-02
#>  [566] 2.276023e-01 9.093051e-02 5.840847e-01 2.805994e-01 7.494842e-01
#>  [571] 7.948750e-01 4.009180e-01 2.536475e-01 3.429767e-01 8.129814e-01
#>  [576] 7.601357e-01 3.148778e-01 9.592469e-01 5.833319e-01 2.652770e-01
#>  [581] 2.219438e-01 9.245375e-01 4.262057e-01 1.630280e-01 9.410540e-01
#>  [586] 7.744225e-01 7.274429e-01 5.580903e-01 8.766353e-03 6.459315e-01
#>  [591] 2.764929e-01 1.294499e-01 6.040567e-01 2.577682e-01 5.204191e-01
#>  [596] 3.081575e-01 7.505825e-01 9.031959e-02 5.839333e-01 3.590320e-01
#>  [601] 1.874069e-01 8.524652e-01 7.411829e-01 1.319659e-01 7.302772e-01
#>  [606] 7.280401e-01 2.674485e-01 7.396386e-01 9.937811e-01 9.158532e-01
#>  [611] 2.232870e-01 6.583251e-01 2.277784e-01 1.099660e-01 9.565532e-01
#>  [616] 3.431873e-01 2.776316e-01 7.077424e-01 6.866407e-01 3.550673e-01
#>  [621] 5.614724e-01 7.994747e-01 1.669566e-01 8.679313e-01 7.465212e-01
#>  [626] 1.003564e-01 8.911707e-02 4.061405e-01 2.867851e-01 1.246053e-01
#>  [631] 7.561402e-01 7.182768e-01 1.520255e-01 1.809036e-01 4.851927e-01
#>  [636] 5.845586e-01 2.063947e-01 8.341374e-01 5.862348e-01 7.336161e-01
#>  [641] 1.374030e-01 1.552088e-01 6.063288e-01 8.881050e-02 6.614613e-01
#>  [646] 1.431773e-01 1.522665e-01 8.191383e-02 3.834797e-01 1.024153e-01
#>  [651] 8.804530e-01 4.254374e-01 7.842035e-01 3.511975e-01 3.494609e-01
#>  [656] 8.933644e-01 8.948860e-01 3.084192e-01 5.520539e-02 7.705742e-01
#>  [661] 4.557562e-01 2.370450e-01 5.866579e-01 9.815024e-01 5.635302e-01
#>  [666] 1.913195e-01 6.683031e-01 2.880332e-01 2.371995e-01 2.315970e-01
#>  [671] 1.993410e-01 1.932464e-01 4.416243e-01 4.558328e-01 7.511110e-01
#>  [676] 2.749461e-01 6.875317e-01 6.128225e-02 5.488885e-01 5.439873e-01
#>  [681] 2.764143e-01 3.475451e-01 3.048136e-01 2.079821e-01 8.466981e-01
#>  [686] 9.827144e-01 4.261723e-01 9.476098e-01 1.275649e-01 2.665210e-01
#>  [691] 8.841565e-01 8.692216e-01 5.900482e-01 4.717168e-01 9.482428e-01
#>  [696] 4.437333e-01 6.280763e-01 1.219550e-01 1.409143e-01 6.966829e-01
#>  [701] 6.145782e-01 2.848213e-01 3.618627e-01 2.321316e-01 4.196723e-01
#>  [706] 6.494500e-01 3.654670e-02 2.642884e-03 4.594015e-01 1.871041e-01
#>  [711] 1.237204e-01 2.172277e-02 5.862153e-01 5.654604e-01 9.932829e-01
#>  [716] 8.778479e-01 1.533280e-01 2.888981e-01 1.125810e-01 2.000958e-02
#>  [721] 3.623547e-02 1.867395e-01 6.006737e-01 9.894985e-01 4.605974e-01
#>  [726] 6.587709e-01 2.150734e-01 3.918479e-01 1.736656e-01 1.165350e-01
#>  [731] 1.806955e-02 3.659735e-01 7.655495e-01 1.575079e-01 6.598505e-01
#>  [736] 9.331903e-01 4.297725e-01 7.280043e-01 1.573318e-01 2.461837e-01
#>  [741] 5.149574e-01 5.810841e-01 9.618195e-01 8.502707e-01 5.955745e-01
#>  [746] 4.676673e-01 2.109651e-02 6.822403e-01 1.910610e-01 6.879974e-01
#>  [751] 1.417310e-01 4.184667e-01 8.748336e-01 5.160320e-01 7.127919e-01
#>  [756] 1.862157e-01 1.671759e-01 2.433068e-02 1.311244e-02 4.936894e-01
#>  [761] 1.920213e-01 9.428024e-01 7.758707e-01 9.824188e-01 1.203631e-01
#>  [766] 9.528757e-01 4.073304e-01 3.272850e-03 2.256920e-02 4.097970e-01
#>  [771] 1.725599e-01 9.593439e-01 3.480026e-01 7.547643e-01 2.418426e-01
#>  [776] 3.505656e-01 1.779432e-01 6.054327e-01 4.133119e-01 2.426442e-01
#>  [781] 6.122967e-01 4.817739e-01 5.778852e-01 1.751137e-01 4.959224e-01
#>  [786] 4.855521e-01 4.288694e-01 3.272740e-01 7.126763e-01 5.941206e-01
#>  [791] 7.173923e-01 8.479271e-01 2.064803e-01 7.677652e-01 8.408466e-01
#>  [796] 5.262039e-01 6.567142e-01 8.289853e-02 1.150261e-01 5.744440e-01
#>  [801] 1.078779e-01 3.645713e-01 1.919978e-01 5.945217e-01 9.356417e-01
#>  [806] 1.751562e-01 5.315525e-01 4.630100e-02 7.443600e-01 9.536151e-01
#>  [811] 9.886818e-01 6.377355e-01 6.354429e-01 9.632055e-01 3.359300e-01
#>  [816] 9.341318e-01 5.478250e-01 1.377514e-01 6.934627e-01 4.422727e-01
#>  [821] 2.779436e-01 2.003882e-01 4.374240e-01 2.722941e-01 3.420865e-01
#>  [826] 1.963684e-01 6.465580e-01 3.878812e-01 8.827484e-01 8.984859e-02
#>  [831] 4.495634e-01 4.791352e-01 2.018548e-01 9.893510e-01 6.971420e-01
#>  [836] 1.480124e-01 7.080017e-01 9.599538e-01 4.051585e-01 1.035957e-01
#>  [841] 7.152330e-01 2.236661e-01 4.444888e-01 3.434944e-01 2.536357e-01
#>  [846] 5.178003e-01 2.484090e-01 2.815200e-01 5.262293e-01 4.090020e-01
#>  [851] 5.698297e-05 7.668141e-03 3.975573e-01 3.351346e-01 4.022110e-01
#>  [856] 6.165786e-01 5.729289e-01 5.933223e-01 2.306403e-01 1.442407e-01
#>  [861] 2.009472e-01 4.943279e-01 1.005442e-01 4.408993e-01 1.320863e-01
#>  [866] 6.449211e-01 9.850494e-01 6.296212e-01 6.100809e-01 6.300376e-01
#>  [871] 6.450246e-01 7.049011e-01 3.018348e-01 6.925567e-01 9.512587e-01
#>  [876] 2.518367e-02 1.444711e-01 5.349564e-01 1.018160e-01 7.005942e-01
#>  [881] 9.361693e-01 4.197982e-01 5.722439e-01 8.617896e-01 8.946732e-01
#>  [886] 4.492323e-01 4.751024e-02 5.285338e-01 3.755010e-03 1.753946e-01
#>  [891] 9.503928e-01 9.425190e-01 5.091777e-01 1.793649e-01 9.564283e-01
#>  [896] 6.000894e-01 5.678489e-03 4.402577e-01 4.999100e-01 7.901656e-01
#>  [901] 8.285700e-01 6.220512e-01 7.525185e-01 9.808566e-01 3.031627e-01
#>  [906] 3.638969e-01 5.737774e-01 7.974222e-01 1.927143e-01 7.096579e-02
#>  [911] 6.637665e-01 5.325533e-01 2.753842e-01 1.657288e-02 8.505290e-01
#>  [916] 3.544237e-01 8.951357e-01 7.594585e-01 7.929092e-01 3.690021e-01
#>  [921] 9.885443e-01 5.202596e-01 4.816507e-01 5.603812e-01 4.059491e-01
#>  [926] 1.780772e-01 2.555985e-01 9.527670e-01 9.492342e-01 2.503465e-01
#>  [931] 4.463720e-01 2.528919e-01 1.102649e-01 1.164325e-01 4.280339e-01
#>  [936] 1.354578e-01 9.509876e-01 5.564700e-01 5.596629e-01 4.858069e-01
#>  [941] 7.061451e-01 4.799485e-01 1.143938e-01 5.256956e-01 4.923019e-01
#>  [946] 6.756639e-01 8.019810e-01 5.067972e-02 2.507613e-01 4.069437e-01
#>  [951] 3.276303e-01 2.237825e-01 9.002859e-01 7.538226e-01 2.925664e-01
#>  [956] 2.633689e-01 4.925530e-01 5.501800e-01 4.931809e-01 2.187330e-01
#>  [961] 2.136674e-01 1.284627e-01 3.190255e-02 7.759703e-01 6.836006e-01
#>  [966] 4.023029e-02 5.471501e-01 9.166965e-01 6.651135e-01 2.377981e-01
#>  [971] 7.333021e-01 2.663977e-01 1.874515e-01 1.774807e-01 2.519407e-01
#>  [976] 7.279319e-01 2.723705e-01 5.014268e-01 6.650865e-01 2.305334e-01
#>  [981] 1.905749e-01 9.799571e-01 2.735355e-01 4.756708e-01 9.019580e-01
#>  [986] 9.228311e-01 3.872787e-01 6.981417e-01 6.761298e-01 7.024149e-01
#>  [991] 5.735192e-01 8.711005e-01 4.285600e-01 4.692609e-01 4.832521e-01
#>  [996] 3.575448e-01 3.148074e-01 6.405638e-01 2.450572e-01 5.444350e-01
#> [1001] 1.345300e-01 8.788860e-01 6.500524e-01 4.292569e-01 3.451318e-01
#> [1006] 7.908153e-01 5.535534e-01 4.987910e-02 3.291545e-01 2.536093e-02
#> [1011] 9.957070e-01 8.555239e-01 1.408815e-01 4.431854e-01 2.354612e-01
#> [1016] 6.397074e-01 4.949633e-01 2.505278e-02 4.825873e-01 9.331015e-01
#> [1021] 4.125940e-01 9.420214e-01 3.686829e-01 9.539024e-01 1.055071e-01
#> [1026] 4.296699e-01 2.853285e-01 5.591850e-01 2.830951e-01 6.670140e-01
#> [1031] 2.007128e-01 9.397887e-01 1.203406e-01 4.366358e-01 3.730674e-01
#> [1036] 2.620464e-01 3.042087e-01 8.369897e-02 1.996781e-01 6.088287e-01
#> [1041] 4.238849e-02 3.726618e-01 8.133340e-02 6.947755e-01 9.492405e-01
#> [1046] 4.260592e-01 3.987327e-01 3.779399e-01 7.706273e-01 1.057720e-01
#> [1051] 4.523877e-01 5.531511e-01 7.217947e-01 4.553711e-01 4.735801e-01
#> [1056] 7.176209e-01 8.148502e-01 8.167231e-01 7.890259e-01 9.493617e-01
#> [1061] 2.648982e-01 8.503215e-01 1.677853e-01 6.148393e-01 8.903800e-01
#> [1066] 3.688923e-01 1.257008e-01 7.099451e-01 6.389231e-01 8.532737e-01
#> [1071] 2.958283e-01 2.291272e-01 4.003309e-01 6.055816e-01 3.764217e-01
#> [1076] 5.560982e-01 6.447780e-01 7.082342e-01 5.132301e-01 1.257772e-01
#> [1081] 6.029584e-01 8.348644e-01 3.884209e-01 1.178496e-01 8.579102e-02
#> [1086] 4.602041e-01 9.649266e-01 5.448183e-02 8.509911e-01 4.747059e-01
#> [1091] 8.797902e-01 1.578948e-01 9.263599e-01 7.144946e-01 3.466359e-01
#> [1096] 2.263795e-02 7.028324e-01 2.026552e-01 6.876447e-01 6.461817e-01
#> [1101] 2.402308e-02 1.257378e-01 4.392930e-01 5.703774e-01 6.898211e-01
#> [1106] 7.362020e-01 3.174739e-01 6.056084e-01 6.166393e-01 3.041729e-01
#> [1111] 2.765572e-01 5.816801e-01 7.028831e-01 1.089807e-01 1.806387e-01
#> [1116] 7.162778e-01 4.613546e-01 8.290836e-02 1.161752e-01 4.343853e-01
#> [1121] 4.051397e-01 9.699104e-01 4.344835e-01 1.131182e-01 6.784924e-01
#> [1126] 7.669101e-01 6.090446e-01 1.205942e-01 7.493839e-01 8.279801e-01
#> [1131] 8.581206e-01 1.756258e-01 4.064683e-01 3.443180e-01 2.782962e-01
#> [1136] 2.433467e-01 6.704846e-01 1.636397e-01 7.818128e-01 6.331869e-01
#> [1141] 7.545399e-01 5.644823e-01 2.052246e-01 8.996652e-01 4.418509e-01
#> [1146] 5.989642e-02 7.166131e-01 2.904225e-01 8.021183e-01 2.575020e-01
#> [1151] 1.338957e-01 6.692374e-01 4.720811e-02 6.985869e-01 9.960922e-01
#> [1156] 9.433979e-01 8.096795e-01 6.483186e-01
#> 
#> $hfunc2
#>    [1] 0.3602979339 0.3868530518 0.3659582292 0.6115207085 0.2979329665
#>    [6] 0.3048357560 0.3438431596 0.6518987137 0.6759897327 0.5547543918
#>   [11] 0.7331852854 0.9573768084 0.3507158679 0.2271521321 0.4020598129
#>   [16] 0.1810535670 0.6723651456 0.8985834539 0.2899298765 0.7204078959
#>   [21] 0.8501991653 0.4603316335 0.4192033381 0.9035938396 0.3615789034
#>   [26] 0.8251909389 0.9015912998 0.4367673416 0.7768123279 0.6203341800
#>   [31] 0.8596998393 0.4823129859 0.7061346845 0.9661257295 0.2500766630
#>   [36] 0.1228342604 0.3847923381 0.2864047692 0.5591397824 0.5677644633
#>   [41] 0.5679038484 0.2036819050 0.4195141707 0.9417793863 0.9203375621
#>   [46] 0.0169195564 0.4676558520 0.1381660940 0.0955139203 0.2523360718
#>   [51] 0.6005016667 0.1129944966 0.4177456511 0.2676088504 0.3327203723
#>   [56] 0.7950253109 0.8032401035 0.0446320259 0.4423884961 0.1230002250
#>   [61] 0.9146304872 0.8457945510 0.5227954517 0.8208220753 0.9092193046
#>   [66] 0.3542929477 0.4994377786 0.2601389222 0.2787378942 0.0120041378
#>   [71] 0.0327136044 0.1643138824 0.2395915278 0.9395575810 0.5173032062
#>   [76] 0.1789059161 0.6667914257 0.6751221222 0.5117736571 0.0528298852
#>   [81] 0.6026595227 0.0047418302 0.7309208537 0.6984327991 0.4686162333
#>   [86] 0.7567127664 0.7501808344 0.4501390535 0.7582690871 0.4885938070
#>   [91] 0.5893851583 0.0667709599 0.2880634034 0.2563626521 0.9402382812
#>   [96] 0.6569588202 0.6650528986 0.6989046736 0.8506293968 0.5476785717
#>  [101] 0.2823542360 0.5648867364 0.6595557701 0.1777972296 0.9509426855
#>  [106] 0.8269240666 0.4160782714 0.7448737647 0.6480376966 0.5092542214
#>  [111] 0.3390633400 0.7734589863 0.6329463011 0.4977176240 0.2863894441
#>  [116] 0.6835141830 0.6851087313 0.2440155461 0.3249232074 0.1906817549
#>  [121] 0.6674449639 0.0407809497 0.6825629177 0.8988593743 0.0464068345
#>  [126] 0.6778086462 0.8534700941 0.3786195945 0.1017327545 0.0491179839
#>  [131] 0.2155701292 0.6155730420 0.8210330787 0.4902426466 0.7756693928
#>  [136] 0.8551385252 0.2981655266 0.4974218833 0.9535755396 0.3808937537
#>  [141] 0.7188332162 0.2160509561 0.3543044306 0.2367622555 0.1178228716
#>  [146] 0.9681445209 0.5775578047 0.4183000484 0.8888511047 0.4590290731
#>  [151] 0.1443458689 0.2093322480 0.5906135398 0.5555647895 0.1265074870
#>  [156] 0.5979861729 0.1673470395 0.3761195295 0.0162187976 0.8828100773
#>  [161] 0.2735453103 0.8127120867 0.6193409711 0.5659049280 0.4306097010
#>  [166] 0.2860994456 0.1896105425 0.7896646138 0.3843820601 0.4332293785
#>  [171] 0.9613750039 0.8413304282 0.6767389183 0.7500778055 0.5829413150
#>  [176] 0.4103787344 0.5949955037 0.9948018742 0.0577988335 0.4237377215
#>  [181] 0.2272214490 0.8628283510 0.3328963264 0.7243015170 0.0577289559
#>  [186] 0.2905064587 0.7226484719 0.9433995271 0.2371001645 0.5947713631
#>  [191] 0.3695327696 0.8349600403 0.9509226820 0.6393767843 0.2643275335
#>  [196] 0.2419421591 0.3649591836 0.4487147695 0.1403573260 0.1071021698
#>  [201] 0.3626918819 0.2446407682 0.0667739427 0.5316676562 0.2297615266
#>  [206] 0.4496800073 0.1249245349 0.9329577781 0.0436974010 0.6279426198
#>  [211] 0.1071748278 0.7266188763 0.7805709719 0.2974466657 0.0904944480
#>  [216] 0.5514406179 0.2853423208 0.8067490394 0.3206238836 0.8537749263
#>  [221] 0.0848937579 0.9946752463 0.6321050810 0.7148637608 0.1642317572
#>  [226] 0.4619152540 0.2748352932 0.8794597591 0.5396498932 0.7253746363
#>  [231] 0.6154912821 0.4326914907 0.1751363451 0.4706118920 0.5470187902
#>  [236] 0.8112137096 0.2099073035 0.0863107290 0.5831665383 0.3950011504
#>  [241] 0.7776498489 0.4876754553 0.1286191136 0.5101264691 0.8045023610
#>  [246] 0.6486328274 0.9398062316 0.1018205001 0.7460307181 0.7348691560
#>  [251] 0.5064798715 0.5169620235 0.4256382861 0.5292969747 0.7308173507
#>  [256] 0.2048741271 0.3616677608 0.0050638761 0.0013326298 0.7708307269
#>  [261] 0.1008532406 0.2770715943 0.3850523754 0.8608938053 0.7669598258
#>  [266] 0.3673760390 0.8952888963 0.8037299760 0.6963857621 0.0562156324
#>  [271] 0.1028489610 0.5774010285 0.5501989469 0.7374746850 0.1182938303
#>  [276] 0.9614141163 0.1420394930 0.9741156100 0.9327547295 0.2123848575
#>  [281] 0.4352987534 0.3102848339 0.3916455436 0.1420868061 0.7015310820
#>  [286] 0.1344012104 0.4547926493 0.6884877611 0.1483746384 0.2428354218
#>  [291] 0.0851547444 0.5644857563 0.7232497365 0.3266771371 0.2715090207
#>  [296] 0.7796778607 0.7087958528 0.1554253588 0.5784324259 0.4653117013
#>  [301] 0.0878993633 0.0803192335 0.9571207312 0.7930953199 0.5906062147
#>  [306] 0.4769853523 0.0685989330 0.4300242324 0.4659937234 0.9702490197
#>  [311] 0.0758323256 0.9187733204 0.1486959878 0.9215475999 0.7052425570
#>  [316] 0.9980161456 0.0431242355 0.2569793271 0.2245581775 0.7668679047
#>  [321] 0.3501016136 0.5912576939 0.2701534475 0.1818662875 0.7835187379
#>  [326] 0.2856799446 0.7783974304 0.4724987686 0.4432145800 0.9102631696
#>  [331] 0.4455895222 0.9397466013 0.9882458539 0.9965969183 0.7839831934
#>  [336] 0.1556048523 0.6392752197 0.0493086385 0.5654789962 0.7010460895
#>  [341] 0.2069559137 0.9809612969 0.0349085349 0.0779248391 0.8713032278
#>  [346] 0.1850185587 0.6872387354 0.1630232149 0.3177454077 0.4227788509
#>  [351] 0.1744492535 0.5745326282 0.6302465667 0.1592283391 0.8014001460
#>  [356] 0.4614678591 0.8945579853 0.5654828447 0.3873400426 0.1894393848
#>  [361] 0.4233942424 0.4061521551 0.4731583391 0.2820430185 0.5458941837
#>  [366] 0.6799183279 0.0995542211 0.3701869870 0.1297803734 0.3119399001
#>  [371] 0.5314272674 0.1506127888 0.7431659566 0.6691214186 0.4843078337
#>  [376] 0.7225475275 0.3855944359 0.1982962225 0.3013871560 0.9563619365
#>  [381] 0.8879615470 0.7334661063 0.6220295111 0.1212285255 0.4902726218
#>  [386] 0.4911635004 0.7329098237 0.3213473728 0.7586615109 0.2600967959
#>  [391] 0.1687731467 0.5446438317 0.5109694447 0.6705053239 0.4377042097
#>  [396] 0.2190664394 0.8098991513 0.5728313548 0.7429565783 0.4427041490
#>  [401] 0.1480158293 0.7887705839 0.2818338689 0.2159913462 0.2720364551
#>  [406] 0.2815133582 0.0797820683 0.4188362262 0.8713283209 0.0835803022
#>  [411] 0.3557827744 0.5080137997 0.8370868197 0.5705359119 0.7693541461
#>  [416] 0.6043845820 0.2719652378 0.6842337032 0.2334529900 0.7640014832
#>  [421] 0.7563851005 0.7042253996 0.3500132756 0.9223013211 0.1158699656
#>  [426] 0.6185513190 0.4981221606 0.3549103254 0.0448196200 0.3377840055
#>  [431] 0.6026228025 0.1379617594 0.3555165811 0.4848435932 0.0601014498
#>  [436] 0.3975567071 0.2979412140 0.5249596146 0.8983687274 0.1834426539
#>  [441] 0.3712017072 0.6347900718 0.6578412735 0.3052217775 0.6886412094
#>  [446] 0.9057192566 0.3897147760 0.6710112965 0.4495019431 0.4177674849
#>  [451] 0.8187897559 0.2593676213 0.6791004975 0.2099329507 0.5852895011
#>  [456] 0.3430087436 0.9454782615 0.7057257494 0.6387367497 0.7697376988
#>  [461] 0.4621322739 0.6640648067 0.9306947637 0.8465153269 0.5031780701
#>  [466] 0.0780049660 0.1904604244 0.2499423742 0.4357465304 0.4410980320
#>  [471] 0.5744816258 0.3849706033 0.9958558135 0.2809152672 0.8995707277
#>  [476] 0.6082616490 0.4515398378 0.6267765211 0.0474084928 0.7406932068
#>  [481] 0.2650152900 0.9935728819 0.3760378594 0.2138039256 0.5045585897
#>  [486] 0.6024735204 0.2819947243 0.2696117674 0.6972424435 0.7310827547
#>  [491] 0.5271869229 0.5261184353 0.7807975922 0.4588799159 0.6541953664
#>  [496] 0.4874196188 0.2331075190 0.9091879778 0.1478415459 0.2721218344
#>  [501] 0.9593452451 0.6445315895 0.2570972752 0.9212444825 0.5562163780
#>  [506] 0.3968182416 0.7999913234 0.1651185324 0.0197905910 0.4681236149
#>  [511] 0.3186847983 0.9174216138 0.1190689090 0.5701578212 0.8195153837
#>  [516] 0.6314311538 0.6079594275 0.6910700713 0.3425548365 0.5852047394
#>  [521] 0.2371454897 0.2494527527 0.5702427898 0.1502267298 0.7660293524
#>  [526] 0.6484564570 0.4440694789 0.9315106979 0.9820654229 0.3250201945
#>  [531] 0.3580767260 0.9564966144 0.3361563328 0.2837581489 0.8207083975
#>  [536] 0.5099573514 0.5481498417 0.4150289830 0.3323853001 0.2632225695
#>  [541] 0.4773000822 0.1419845900 0.9917787986 0.1713372742 0.4012782196
#>  [546] 0.2283306405 0.1712470170 0.6998318501 0.6689217315 0.3366585054
#>  [551] 0.7726896682 0.8373391332 0.5326914955 0.6122537637 0.6742242959
#>  [556] 0.1451321126 0.2067799555 0.8819340022 0.3022965728 0.6843682831
#>  [561] 0.7073954719 0.5808120353 0.5828749702 0.3505242929 0.9206796824
#>  [566] 0.9565310104 0.9211499208 0.8013813942 0.7440514274 0.3173967289
#>  [571] 0.5934856556 0.7875299366 0.8627221595 0.8443053733 0.1017885647
#>  [576] 0.2031863013 0.9883196347 0.5881294299 0.2529065440 0.3676572059
#>  [581] 0.5062390399 0.4160138111 0.3694870508 0.5450923117 0.1343427662
#>  [586] 0.1345839783 0.2586427552 0.3399446211 0.9498602200 0.6175089988
#>  [591] 0.6841834814 0.3828409457 0.5144211889 0.3852818459 0.5603676550
#>  [596] 0.6138703117 0.6119633021 0.7530404711 0.4131433503 0.9742692989
#>  [601] 0.9483571101 0.2894445700 0.1300378616 0.6541669915 0.4888190918
#>  [606] 0.2008041341 0.3637642277 0.9442322833 0.1787967663 0.1649434444
#>  [611] 0.6286538402 0.0015834919 0.2518212470 0.6587865058 0.5380582721
#>  [616] 0.5939064961 0.4491298789 0.7835408055 0.9233006198 0.9031842325
#>  [621] 0.5970617738 0.7033117325 0.2852440792 0.2814720887 0.1685589830
#>  [626] 0.4954628965 0.7008137047 0.9279244358 0.9544190104 0.4425845903
#>  [631] 0.6778145472 0.3269640337 0.4932159378 0.9544095654 0.6187518627
#>  [636] 0.0628787703 0.3834242498 0.6092924949 0.8322686591 0.4273459350
#>  [641] 0.7435800262 0.2520850956 0.6524576421 0.4591370391 0.4925247851
#>  [646] 0.2250845442 0.3388182436 0.5161007119 0.0781148254 0.9714643358
#>  [651] 0.6103408978 0.1117112335 0.5063627616 0.1420287424 0.8888880489
#>  [656] 0.3533968613 0.3102264245 0.0786184052 0.7450230880 0.7783252962
#>  [661] 0.1762540205 0.6645075079 0.0891146349 0.1385400393 0.5965295990
#>  [666] 0.9193356789 0.4374766894 0.6810049571 0.6252272225 0.7432358371
#>  [671] 0.5542695684 0.9133417890 0.8731864268 0.7981255798 0.6538209866
#>  [676] 0.0817422474 0.3776380069 0.3047958308 0.1279130689 0.6637773741
#>  [681] 0.5315741689 0.7751062983 0.4770269016 0.7720777393 0.6988783672
#>  [686] 0.5385293745 0.4502497149 0.1572084089 0.9313229601 0.7383589875
#>  [691] 0.0631670850 0.4117803253 0.3852160276 0.5463312029 0.0134753756
#>  [696] 0.4350046018 0.2752307041 0.9098316454 0.9335615102 0.1058721986
#>  [701] 0.1647209640 0.5465460682 0.7126085574 0.3768164001 0.6399556314
#>  [706] 0.6575854322 0.6735478428 0.9897076150 0.1195294778 0.9380343948
#>  [711] 0.3828268882 0.9936176498 0.5356377855 0.6125700280 0.0007258805
#>  [716] 0.1664341622 0.1083716059 0.5326668141 0.6990299332 0.9479788512
#>  [721] 0.8774657694 0.4163648637 0.2464681604 0.0100767851 0.7491120895
#>  [726] 0.6081333149 0.1248321647 0.2581851166 0.0685509851 0.9930060703
#>  [731] 0.9199901938 0.7954498012 0.3425936113 0.4908804482 0.1603113184
#>  [736] 0.7914049681 0.6858096896 0.5085608840 0.6278597448 0.5891867459
#>  [741] 0.8817490070 0.5464230752 0.2644109083 0.1397528062 0.5020074870
#>  [746] 0.8469771376 0.8945431646 0.3057523803 0.3135583532 0.4446068513
#>  [751] 0.8135067220 0.7774050228 0.8859211917 0.6286733542 0.3086519804
#>  [756] 0.4961253740 0.8344396645 0.9219217323 0.9957776294 0.4093742273
#>  [761] 0.8526333125 0.0001640302 0.3490491136 0.0986295370 0.5712219249
#>  [766] 0.0049584063 0.3878337303 0.4362869785 0.0519495000 0.6071585400
#>  [771] 0.1079214016 0.6458289905 0.5431988616 0.2268263762 0.9054858902
#>  [776] 0.5040338242 0.8538869019 0.5469740965 0.4709729113 0.2379225277
#>  [781] 0.5668521191 0.2233239624 0.4110733883 0.6896923724 0.8758725332
#>  [786] 0.5461013774 0.6253623280 0.2328972723 0.6860842092 0.7066716245
#>  [791] 0.1343031000 0.0746949083 0.4545291504 0.4593686761 0.6682778223
#>  [796] 0.4537520604 0.1322659599 0.4269183921 0.2398109065 0.9364632545
#>  [801] 0.6105752293 0.2954731387 0.4752145570 0.8387753714 0.2306671937
#>  [806] 0.5414366051 0.3745781024 0.9848598457 0.2294966299 0.0293642971
#>  [811] 0.5273489347 0.1067626919 0.8431006234 0.0535251226 0.6747745621
#>  [816] 0.6613380361 0.7923559965 0.7454470104 0.3924333155 0.6188390817
#>  [821] 0.6308133565 0.5610336005 0.7780190800 0.2902543493 0.3936265088
#>  [826] 0.9503247082 0.8077370533 0.5853026462 0.7706971858 0.6064350805
#>  [831] 0.1860955249 0.8000675621 0.8528943643 0.0002006685 0.4672945157
#>  [836] 0.7522562741 0.6114568294 0.0772057037 0.7087419993 0.6065628841
#>  [841] 0.2590805111 0.5261388642 0.1423774287 0.7584739633 0.9302268168
#>  [846] 0.4490237304 0.7318116970 0.8915450119 0.6369933654 0.4216906145
#>  [851] 0.9976182685 0.8850415373 0.2327857374 0.7940400213 0.4140635150
#>  [856] 0.8561550590 0.5109281879 0.7739211740 0.2249269682 0.9599012384
#>  [861] 0.6379893604 0.3214197491 0.1662563554 0.3644012959 0.3459957385
#>  [866] 0.7929167368 0.2740586199 0.0849448768 0.7155296516 0.0556240379
#>  [871] 0.3186948700 0.0158202228 0.7784126532 0.2455197531 0.1164226152
#>  [876] 0.5236479456 0.5773258872 0.2702102229 0.3890843112 0.1029379914
#>  [881] 0.1342264223 0.2403077970 0.7915686118 0.0197189976 0.0951981056
#>  [886] 0.3257221502 0.3520683945 0.5563018717 0.7483001523 0.7506170572
#>  [891] 0.4802876791 0.4130999684 0.6216935120 0.9534252759 0.1948125069
#>  [896] 0.0511191561 0.9799752580 0.1637057785 0.5952557538 0.1615089494
#>  [901] 0.0763169687 0.1085469257 0.0962676082 0.2672919289 0.6089525763
#>  [906] 0.4913850119 0.4533114589 0.4447804629 0.8357597600 0.4005345779
#>  [911] 0.3598711731 0.3561619335 0.7444995604 0.7305090066 0.1698977444
#>  [916] 0.1497214184 0.4487057125 0.1428669277 0.5196746762 0.4128227309
#>  [921] 0.0659276833 0.6041693274 0.5205833782 0.8817540999 0.3104817764
#>  [926] 0.0996294498 0.0636116006 0.4134992298 0.0103609680 0.3736973134
#>  [931] 0.4790275062 0.8637950973 0.0311183938 0.1821853932 0.0553248371
#>  [936] 0.2379918657 0.8965856992 0.2075600333 0.5378059980 0.5657839434
#>  [941] 0.6759399069 0.3616786157 0.2011128865 0.5245465558 0.4633659923
#>  [946] 0.0520071299 0.6643116194 0.1949919159 0.4217891441 0.0076760508
#>  [951] 0.6238235883 0.0384178366 0.8728699010 0.5099796345 0.0961255861
#>  [956] 0.1404439938 0.6354482109 0.5397277017 0.5551195377 0.3162814086
#>  [961] 0.5419412871 0.3793455795 0.4022199144 0.9706209384 0.2856490825
#>  [966] 0.9762631572 0.7606160280 0.5626717376 0.1863443606 0.1552142899
#>  [971] 0.5242510271 0.8638923434 0.3721061163 0.4841737192 0.7038044054
#>  [976] 0.4189399800 0.5164203491 0.4547907977 0.0051510140 0.7590672425
#>  [981] 0.4434466707 0.7273803642 0.8077522359 0.3385753969 0.4097195083
#>  [986] 0.2235144778 0.1253816403 0.5574280294 0.2150598766 0.5458425378
#>  [991] 0.1668320943 0.8413062845 0.7423587766 0.7456963661 0.3129147954
#>  [996] 0.3742409623 0.6650375876 0.7711914404 0.6415724541 0.5602540147
#> [1001] 0.6188156846 0.2124016734 0.3089051399 0.8485943190 0.8742153591
#> [1006] 0.4067439310 0.3212779474 0.7480535795 0.1424894399 0.8775738500
#> [1011] 0.0051011732 0.2212391152 0.6238164749 0.0353481381 0.2044880829
#> [1016] 0.7185622324 0.0788015927 0.8824484977 0.5892186399 0.1306650442
#> [1021] 0.1007365260 0.5049888087 0.4022016401 0.5823915514 0.9579785082
#> [1026] 0.1117105292 0.3377216567 0.6468134720 0.9915249303 0.3139406065
#> [1031] 0.9910919139 0.2687688896 0.2636426345 0.4997814528 0.2095245811
#> [1036] 0.8092740227 0.5116563266 0.2562519572 0.6419063783 0.3062791384
#> [1041] 0.2612783727 0.4045980680 0.8530145561 0.1153688835 0.8506425013
#> [1046] 0.2435012044 0.1485147123 0.4127696220 0.9089896869 0.1934780493
#> [1051] 0.4908700485 0.4797994594 0.9430511917 0.8263137876 0.7052891438
#> [1056] 0.3555770606 0.3382903796 0.2013632703 0.1899537720 0.2350227101
#> [1061] 0.7463738984 0.5877167359 0.7747166861 0.2142283148 0.2290652929
#> [1066] 0.2860088308 0.0689789055 0.6619440928 0.5744474503 0.9198727571
#> [1071] 0.9148936865 0.5300326250 0.3395870430 0.4515438576 0.9658781856
#> [1076] 0.7619702302 0.2427006333 0.3199363239 0.5156448205 0.2502346306
#> [1081] 0.3460812149 0.6086284737 0.2214564158 0.9824025897 0.8067799183
#> [1086] 0.2865817951 0.1149667443 0.9977094565 0.4736712233 0.4825431907
#> [1091] 0.3558017876 0.5099751349 0.5246431754 0.2412751788 0.6697760635
#> [1096] 0.2786853033 0.1153124940 0.8486932573 0.7794308369 0.6011647077
#> [1101] 0.9968299241 0.3062592829 0.6249847913 0.4725776085 0.6342648330
#> [1106] 0.3411507271 0.3286367125 0.3321631536 0.9838202770 0.5953493906
#> [1111] 0.5557981485 0.2494124446 0.4023704351 0.5113250692 0.6088485060
#> [1116] 0.5134947308 0.7290613227 0.8074504491 0.2750230223 0.2916297335
#> [1121] 0.0558592780 0.0776608107 0.5502043781 0.1925702617 0.3470351456
#> [1126] 0.8257441805 0.1111294624 0.6245933453 0.8866048110 0.0560099033
#> [1131] 0.6895549189 0.0856778460 0.5612822489 0.5368451961 0.5767791342
#> [1136] 0.5954236452 0.5543859131 0.6204617325 0.9345829161 0.7434144567
#> [1141] 0.8912522963 0.6551713275 0.7374149371 0.3752647456 0.8347036744
#> [1146] 0.9201580009 0.8533951659 0.5439155892 0.1892795168 0.5452197446
#> [1151] 0.9640701556 0.1469945355 0.9782907595 0.8144351537 0.0263820822
#> [1156] 0.1861138014 0.1154144500 0.4989536219
#> 
# or using the fast versions
h1 <- BiCopHfunc1(daxreturns[, 2], daxreturns[, 1], cop)
h2 <- BiCopHfunc2(daxreturns[, 2], daxreturns[, 1], cop)
all.equal(h$hfunc1, h1)
#> [1] TRUE
all.equal(h$hfunc2, h2)
#> [1] TRUE