Bivariate Copula Family Names

This function transforms the bivariate copula family number into its character expression and vice versa.

BiCopName(family, short = TRUE)

Arguments

family
Bivariate copula family, either its number or its character expression (see table below).

No.	Short name

Long name		0

"I"

"Independence"

1

"N"

"Gaussian"

2

"t"

"t"

3

"C"

"Clayton"

4

"G"

"Gumbel"

5

"F"

"Frank"

6

"J"

"Joe"

7

"BB1"

"BB1"

8

"BB6"

"BB6"

9

"BB7"

"BB7"

10

"BB8"

"Frank-Joe"

13

"SC"

"Survival Clayton"

14

"SG"

"Survival Gumbel"

16

"SJ"

"Survival Joe"

17

"SBB1"

"Survival BB1"

18

"SBB6"

"Survival BB6"

19

"SBB7"

"Survival BB7"

20

"SBB8"

"Survival BB8"

23

"C90"

"Rotated Clayton 90 degrees"

24

"G90"

"Rotated Gumbel 90 degrees"

26

"J90"

"Rotated Joe 90 degrees"

27

"BB1_90"

"Rotated BB1 90 degrees"

28

"BB6_90"

"Rotated BB6 90 degrees"

29

"BB7_90"

"Rotated BB7 90 degrees"

30

"BB8_90"

"Rotated Frank-Joe 90 degrees"

33

"C270"

"Rotated Clayton 270 degrees"

34

"G270"

"Rotated Gumbel 270 degrees"

36

"J270"

"Rotated Joe 270 degrees"

37

"BB1_270"

"Rotated BB1 270 degrees"

38

"BB6_270"

"Rotated BB6 270 degrees"

39

"BB7_270"

"Rotated BB7 270 degrees"

40

"BB8_270"

"Rotated Frank-Joe 270 degrees"

104

"Tawn"

"Tawn type 1"

114

"Tawn180"

"Rotated Tawn type 1 180 degrees"

124

"Tawn90"

"Rotated Tawn type 1 90 degrees"

134

"Tawn270"

"Rotated Tawn type 1 270 degrees"

204

"Tawn2"

"Tawn type 2"

214

"Tawn2_180"

"Rotated Tawn type 2 180 degrees"

224

"Tawn2_90"

"Rotated Tawn type 2 90 degrees"

234

"Tawn2_270"

"Rotated Tawn type 2 270 degrees"

short
Logical; if the number of a bivariate copula family is used and short = TRUE (default), a short version of the corresponding character expression is returned, otherwise the long version.

Value

The transformed bivariate copula family (see table above).

See also

Examples

## family number to character expression
family <- 1
BiCopName(family, short = TRUE)  # short version

#> [1] "N"

BiCopName(family, short = FALSE)         # long version

#> [1] "Gaussian"

## family character expression (short version) to number
family <- "C"
BiCopName(family)       # as number

#> C 
#> 3 

## family character expression (long version) to number
family <- "Clayton"
BiCopName(family)       # as number

#> Clayton 
#>       3 

## vectors of families
BiCopName(1:10)    # as character expression

#>  [1] "N"   "t"   "C"   "G"   "F"   "J"   "BB1" "BB6" "BB7" "BB8"

BiCopName(c("Clayton","t","J"))    # as number

#> Clayton       t       J 
#>       3       2       6