Density of a Bivariate Copula

This function evaluates the probability density function (PDF) of a given parametric bivariate copula.

BiCopPDF(u1, u2, family, par, par2 = 0, obj = NULL, check.pars = TRUE)

Arguments

u1, u2
numeric vectors of equal length with values in [0,1].

family
integer; single number or vector of size length(u1); defines the bivariate copula family: 0 = independence copula 1 = Gaussian copula 2 = Student t copula (t-copula) 3 = Clayton copula 4 = Gumbel copula 5 = Frank copula 6 = Joe copula 7 = BB1 copula 8 = BB6 copula 9 = BB7 copula 10 = BB8 copula 13 = rotated Clayton copula (180 degrees; ``survival Clayton'') 14 = rotated Gumbel copula (180 degrees; ``survival Gumbel'') 16 = rotated Joe copula (180 degrees; ``survival Joe'') 17 = rotated BB1 copula (180 degrees; ``survival BB1'') 18 = rotated BB6 copula (180 degrees; ``survival BB6'') 19 = rotated BB7 copula (180 degrees; ``survival BB7'') 20 = rotated BB8 copula (180 degrees; ``survival BB8'') 23 = rotated Clayton copula (90 degrees) 24 = rotated Gumbel copula (90 degrees) 26 = rotated Joe copula (90 degrees) 27 = rotated BB1 copula (90 degrees) 28 = rotated BB6 copula (90 degrees) 29 = rotated BB7 copula (90 degrees) 30 = rotated BB8 copula (90 degrees) 33 = rotated Clayton copula (270 degrees) 34 = rotated Gumbel copula (270 degrees) 36 = rotated Joe copula (270 degrees) 37 = rotated BB1 copula (270 degrees) 38 = rotated BB6 copula (270 degrees) 39 = rotated BB7 copula (270 degrees) 40 = rotated BB8 copula (270 degrees) 104 = Tawn type 1 copula 114 = rotated Tawn type 1 copula (180 degrees) 124 = rotated Tawn type 1 copula (90 degrees) 134 = rotated Tawn type 1 copula (270 degrees) 204 = Tawn type 2 copula 214 = rotated Tawn type 2 copula (180 degrees) 224 = rotated Tawn type 2 copula (90 degrees) 234 = rotated Tawn type 2 copula (270 degrees)

par
numeric; single number or vector of size length(u1); copula parameter.

par2
numeric; single number or vector of size length(u1); second parameter for bivariate copulas with two parameters (t, BB1, BB6, BB7, BB8, Tawn type 1 and type 2; default: par2 = 0). par2 should be an positive integer for the Students's t copula family = 2.

obj
BiCop object containing the family and parameter specification.

check.pars
logical; default is TRUE; if FALSE, checks for family/parameter-consistency are ommited (should only be used with care).

Value

A numeric vector of the bivariate copula density

of the copula family

with parameter(s) par, par2

evaluated at u1 and u2.

Details

If the family and parameter specification is stored in a BiCop object obj, the alternative version

BiCopPDF(u1, u2, obj)

can be used.

See also

BiCopCDF, BiCopHfunc, BiCopSim, BiCop

Examples

## simulate from a bivariate Student-t copula
cop <- BiCop(family = 2, par = -0.7, par2 = 4)
simdata <- BiCopSim(100, cop)

## evaluate the density of the bivariate t-copula
u1 <- simdata[,1]
u2 <- simdata[,2]
BiCopPDF(u1, u2, cop)

#>   [1]  2.1306456  4.1962716  9.1524595  0.8952852  1.4656171  1.2886303
#>   [7]  0.1538810  1.4331731  1.3010896  0.2505423  1.4469876  0.4073968
#>  [13]  0.7353143  1.4460763  9.1858944  1.8387049  1.6328029  1.1686813
#>  [19]  1.6916863  0.9085196  1.8593422  0.6583833  2.0996375  2.4533072
#>  [25]  2.3866976  8.4577953  1.2008736  0.1797626  1.0483543  2.0665501
#>  [31]  4.4485417  1.6137590  1.8568620  3.5485391  4.8100314  1.5457310
#>  [37]  2.1103129  1.6153957  2.1622530  1.1592634  1.5612059  0.5663747
#>  [43]  1.2653649  1.5913829  1.5946645  2.2793286  2.5445049  2.8946325
#>  [49]  0.2324600  0.7565870  1.5291350  1.0309052  1.0607904  0.5233480
#>  [55]  2.9304414  1.3802993  1.5835675  2.8899306  2.1266217  1.1420443
#>  [61]  1.2639900  1.0639686  2.5038283  1.7124491  1.0572096  1.8474986
#>  [67]  2.9176118  0.3158799  2.0321830  3.1898574  1.7025635  2.8849251
#>  [73]  1.2206931  0.3798045 23.3516742  2.6714202  0.6345198  0.6190335
#>  [79]  1.8902789  7.6958835  0.5790888  1.8217116  7.1691876  7.7269896
#>  [85]  1.0912906  2.7140931  1.5055683  1.2056177  5.5423714  1.5052213
#>  [91]  1.2178262  0.8869690  1.2487186  0.7491671  0.3550704  3.0378246
#>  [97]  1.2558922  1.7871041  0.4751990  1.7346539

## select a bivariate copula for the simulated data
fit <- BiCopSelect(u1, u2)
summary(fit)

#> Family
#> ------ 
#> No:    39
#> Name:  Rotated BB7 270 degrees
#> 
#> Parameter(s)
#> ------------
#> par:  -1.98
#> par2: -1.04
#> Dependence measures
#> -------------------
#> Kendall's tau:    -0.5 (empirical = -0.53, p value < 0.01)
#> Upper TD:         0 
#> Lower TD:         0 
#> 
#> Fit statistics
#> --------------
#> logLik:  44.29 
#> AIC:    -84.58 
#> BIC:    -79.37 
#> 

## and evaluate its PDF
round(BiCopPDF(u1, u2, fit), 3)

#>   [1]  2.043  4.298 10.238  0.996  1.430  1.295  0.118  1.388  1.151  0.234
#>  [11]  1.405  0.464  0.803  1.396 10.228  1.768  1.583  1.207  1.571  1.006
#>  [21]  1.867  0.583  1.926  2.485  2.241  9.409  1.222  0.145  1.115  2.039
#>  [31]  4.790  1.497  1.648  3.690  5.686  1.467  1.966  1.540  2.009  1.198
#>  [41]  1.487  0.607  1.278  1.486  1.502  2.285  2.240  2.948  0.256  0.721
#>  [51]  1.494  1.087  1.128  0.642  2.854  1.365  1.490  2.960  2.150  1.169
#>  [61]  1.294  1.134  2.566  1.584  1.097  1.771  2.840  0.347  1.992  3.506
#>  [71]  1.581  2.854  1.250  0.470 32.092  2.551  0.759  0.723  1.833  9.888
#>  [81]  0.663  1.808  9.442 10.559  1.117  2.745  1.439  1.168  2.297  1.443
#>  [91]  1.245  0.993  1.266  0.857  0.435  3.255  1.271  1.712  0.569  1.650