Tail Dependence Coefficients of a Bivariate Copula

This function computes the theoretical tail dependence coefficients of a bivariate copula for given parameter values.

BiCopPar2TailDep(family, par, par2 = 0, obj = NULL, check.pars = TRUE)

Arguments

family

integer; single number or vector of size n; defines the bivariate copula family:
0 = independence copula
1 = Gaussian copula
2 = Student t copula (t-copula)
3 = Clayton copula
4 = Gumbel copula
5 = Frank copula
6 = Joe copula
7 = BB1 copula
8 = BB6 copula
9 = BB7 copula
10 = BB8 copula
13 = rotated Clayton copula (180 degrees; survival Clayton'') \cr `14` = rotated Gumbel copula (180 degrees; survival Gumbel'')
16 = rotated Joe copula (180 degrees; survival Joe'') \cr `17` = rotated BB1 copula (180 degrees; survival BB1'')
18 = rotated BB6 copula (180 degrees; survival BB6'')\cr `19` = rotated BB7 copula (180 degrees; survival BB7'')
20 = rotated BB8 copula (180 degrees; ``survival BB8'')
23 = rotated Clayton copula (90 degrees)
`24` = rotated Gumbel copula (90 degrees)
`26` = rotated Joe copula (90 degrees)
`27` = rotated BB1 copula (90 degrees)
`28` = rotated BB6 copula (90 degrees)
`29` = rotated BB7 copula (90 degrees)
`30` = rotated BB8 copula (90 degrees)
`33` = rotated Clayton copula (270 degrees)
`34` = rotated Gumbel copula (270 degrees)
`36` = rotated Joe copula (270 degrees)
`37` = rotated BB1 copula (270 degrees)
`38` = rotated BB6 copula (270 degrees)
`39` = rotated BB7 copula (270 degrees)
`40` = rotated BB8 copula (270 degrees)
`104` = Tawn type 1 copula
`114` = rotated Tawn type 1 copula (180 degrees)
`124` = rotated Tawn type 1 copula (90 degrees)
`134` = rotated Tawn type 1 copula (270 degrees)
`204` = Tawn type 2 copula
`214` = rotated Tawn type 2 copula (180 degrees)
`224` = rotated Tawn type 2 copula (90 degrees)
`234` = rotated Tawn type 2 copula (270 degrees)

par

numeric; single number or vector of size n; copula parameter.

par2

numeric; single number or vector of size n; second parameter for bivariate copulas with two parameters (t, BB1, BB6, BB7, BB8, Tawn type 1 and type 2; default: par2 = 0). par2 should be an positive integer for the Students's t copula family = 2.

obj

BiCop object containing the family and parameter specification.

check.pars

logical; default is TRUE; if FALSE, checks for family/parameter-consistency are omitted (should only be used with care).

Value

lower

Lower tail dependence coefficient for the given bivariate copula family and parameter(s) par, par2: $$ \lambda_L = \lim_{u\searrow 0}\frac{C(u,u)}{u} $$

upper

Upper tail dependence coefficient for the given bivariate copula family family and parameter(s) par, par2: $$ \lambda_U = \lim_{u\nearrow 1}\frac{1-2u+C(u,u)}{1-u} $$

Lower and upper tail dependence coefficients for bivariate copula families and parameters (\(\theta\) for one parameter families and the first parameter of the t-copula with \(\nu\) degrees of freedom, \(\theta\) and \(\delta\) for the two parameter BB1, BB6, BB7 and BB8 copulas) are given in the following table.

No.	Lower tail dependence	Upper tail dependence

1	-	-

2	\(2t_{\nu+1}\left(-\sqrt{\nu+1}\sqrt{\frac{1-\theta}{1+\theta}}\right)\)	\(2t_{\nu+1}\left(-\sqrt{\nu+1}\sqrt{\frac{1-\theta}{1+\theta}}\right)\)

3	\(2^{-1/\theta}\)	-

4	-	\(2-2^{1/\theta}\)

5	-	-

6	-	\(2-2^{1/\theta}\)

7	\(2^{-1/(\theta\delta)}\)	\(2-2^{1/\delta}\)

8	-	\(2-2^{1/(\theta\delta)}\)

9	\(2^{-1/\delta}\)	\(2-2^{1/\theta}\)

10	-	\(2-2^{1/\theta}\) if \(\delta=1\) otherwise 0

13	-	\(2^{-1/\theta}\)

14	\(2-2^{1/\theta}\)	-

16	\(2-2^{1/\theta}\)	-

17	\(2-2^{1/\delta}\)	\(2^{-1/(\theta\delta)}\)

18	\(2-2^{1/(\theta\delta)}\)	-

19	\(2-2^{1/\theta}\)	\(2^{-1/\delta}\)

20	\(2-2^{1/\theta}\) if \(\delta=1\) otherwise 0	-

23, 33	-	-

24, 34	-	-

26, 36	-	-

27, 37	-	-

28, 38	-	-

29, 39	-	-

30, 40	-	-

104,204	-	\(\delta+1-(\delta^{\theta}+1)^{1/\theta}\)

114, 214	\(1+\delta-(\delta^{\theta}+1)^{1/\theta}\)	-

124, 224	-	-

134, 234	-	-

Details

If the family and parameter specification is stored in a BiCop object obj, the alternative version

BiCopPar2TailDep(obj)

can be used.

Note

The number n can be chosen arbitrarily, but must agree across arguments.

References

Joe, H. (1997). Multivariate Models and Dependence Concepts. Chapman and Hall, London.

See also

Author

Eike Brechmann

Examples

## Example 1: Gaussian copula
BiCopPar2TailDep(1, 0.7)
#> $lower
#> [1] 0
#> 
#> $upper
#> [1] 0
#> 
BiCop(1, 0.7)$taildep  # alternative
#> $lower
#> [1] 0
#> 
#> $upper
#> [1] 0
#> 

## Example 2: Student-t copula
BiCopPar2TailDep(2, c(0.6, 0.7, 0.8), 4)
#> $lower
#> [1] 0.3143726 0.3906840 0.4895897
#> 
#> $upper
#> [1] 0.3143726 0.3906840 0.4895897
#> 

## Example 3: different copula families
BiCopPar2TailDep(c(3, 4, 6), 2)
#> $lower
#> [1] 0.7071068 0.0000000 0.0000000
#> 
#> $upper
#> [1] 0.0000000 0.5857864 0.5857864
#>