Randomv sampling of R-Vine matices
Source: R/RVineMatrixSample.R

Sample R-Vine matrices based on the algorithm of Joe et al. (2011).

RVineMatrixSample(d, size = 1, naturalOrder = FALSE)

Arguments

d
Dimension of the R-Vine matrices.

size
Number of matrices to sample.

naturalOrder
Should the matrices be in the natural order (default: naturalOrder = FALSE).

Value

A list of length size with each element containing one R-Vine matrix.

Note

For some reason, our implementaion of Joe et al.'s algorithm always returns a star in the first tree. To fix this, we sample a vine matrix of dimension d + 1 and remove the first tree afterwards

References

Joe H, Cooke RM and Kurowicka D (2011). Regular vines: generation algorithm and number of equivalence classes. In Dependence Modeling: Vine Copula Handbook, pp 219--231. World Scientific, Singapore.

See also

Examples

# Matrix and sample sizes
d <- 10
size <- 5

# Sample R-vine matrices
RVM <- RVineMatrixSample(d, size)
sapply(RVM, RVineMatrixCheck)

#> [1] 1 1 1 1 1

# Sample R-vine matrices in the natural order
RVM <- RVineMatrixSample(d, size, naturalOrder = TRUE)
sapply(RVM, RVineMatrixCheck)

#> [1] 1 1 1 1 1