Copula Data Objects

The function as.copuladata coerces an object (data.frame, matrix, list) to a copuladata object.

as.copuladata(data)

Arguments

data

Either a data.frame, a matrix or a list containing copula data (i.e. data with uniform margins on \([0,1]\)). The list elements have to be vectors of identical length.

See also

Author

Tobias Erhardt

Examples

data(daxreturns)

data <- as.matrix(daxreturns)
class(as.copuladata(data))
#> [1] "copuladata" "data.frame"

data <- as.data.frame(daxreturns)
class(as.copuladata(data))
#> [1] "copuladata" "data.frame"

data <- as.list(daxreturns)
names(data) <- names(daxreturns)
class(as.copuladata(data))
#> [1] "copuladata" "data.frame"