Distribution Function of a Bivariate Copula

This function evaluates the cumulative distribution function (CDF) of a given parametric bivariate copula.

BiCopCDF(u1, u2, family, par, par2 = 0, obj = NULL, check.pars = TRUE)

Arguments

u1, u2
numeric vectors of equal length with values in \([0,1]\).

family
integer; single number or vector of size length(u1); defines the bivariate copula family:
0 = independence copula
1 = Gaussian copula
2 = Student t copula (t-copula)
3 = Clayton copula
4 = Gumbel copula
5 = Frank copula
6 = Joe copula
7 = BB1 copula
8 = BB6 copula
9 = BB7 copula
10 = BB8 copula
13 = rotated Clayton copula (180 degrees; survival Clayton'') \cr `14` = rotated Gumbel copula (180 degrees; survival Gumbel'')
16 = rotated Joe copula (180 degrees; survival Joe'') \cr `17` = rotated BB1 copula (180 degrees; survival BB1'')
18 = rotated BB6 copula (180 degrees; survival BB6'')\cr `19` = rotated BB7 copula (180 degrees; survival BB7'')
20 = rotated BB8 copula (180 degrees; ``survival BB8'')
23 = rotated Clayton copula (90 degrees)
`24` = rotated Gumbel copula (90 degrees)
`26` = rotated Joe copula (90 degrees)
`27` = rotated BB1 copula (90 degrees)
`28` = rotated BB6 copula (90 degrees)
`29` = rotated BB7 copula (90 degrees)
`30` = rotated BB8 copula (90 degrees)
`33` = rotated Clayton copula (270 degrees)
`34` = rotated Gumbel copula (270 degrees)
`36` = rotated Joe copula (270 degrees)
`37` = rotated BB1 copula (270 degrees)
`38` = rotated BB6 copula (270 degrees)
`39` = rotated BB7 copula (270 degrees)
`40` = rotated BB8 copula (270 degrees)
`104` = Tawn type 1 copula
`114` = rotated Tawn type 1 copula (180 degrees)
`124` = rotated Tawn type 1 copula (90 degrees)
`134` = rotated Tawn type 1 copula (270 degrees)
`204` = Tawn type 2 copula
`214` = rotated Tawn type 2 copula (180 degrees)
`224` = rotated Tawn type 2 copula (90 degrees)
`234` = rotated Tawn type 2 copula (270 degrees)

par
numeric; single number or vector of size length(u1); copula parameter.

par2
numeric; single number or vector of size length(u1); second parameter for bivariate copulas with two parameters (BB1, BB6, BB7, BB8, Tawn type 1 and type 2; default: par2 = 0).

obj
BiCop object containing the family and parameter specification.

check.pars
logical; default is TRUE; if FALSE, checks for family/parameter-consistency are omitted (should only be used with care).

Value

A numeric vector of the bivariate copula distribution function

of the copula family

with parameter(s) par, par2

evaluated at u1 and u2.

Details

If the family and parameter specification is stored in a BiCop() object obj, the alternative version

BiCopCDF(u1, u2, obj)

can be used.

Note

The calculation of the cumulative distribution function (CDF) of the Student's t copula (family = 2) is only approximate. For numerical reasons, the degree of freedom parameter (par2) is rounded to an integer before calculation of the CDF.

See also

Examples

## simulate from a bivariate Clayton copula
set.seed(123)
cop <- BiCop(family = 3, par = 3.4)
simdata <- BiCopSim(300, cop)

## evaluate the distribution function of the bivariate Clayton copula
u1 <- simdata[,1]
u2 <- simdata[,2]
BiCopCDF(u1, u2, cop)

#>   [1] 0.272443455 0.397575020 0.466088911 0.514619601 0.461444338 0.799373201
#>   [7] 0.596934319 0.100484924 0.119765543 0.324468736 0.818352006 0.639670814
#>  [13] 0.606316797 0.483351456 0.187054448 0.940783304 0.655702676 0.020810245
#>  [19] 0.535621671 0.228195797 0.116899611 0.329812119 0.097320715 0.195904623
#>  [25] 0.256862448 0.038073231 0.495198018 0.391956261 0.119579624 0.715966111
#>  [31] 0.389399721 0.286187740 0.678929140 0.772700884 0.659077453 0.679076063
#>  [37] 0.132777661 0.336966627 0.339807182 0.213521471 0.222278103 0.395655216
#>  [43] 0.085129534 0.959959230 0.596563210 0.118634633 0.312211436 0.219043218
#>  [49] 0.456629470 0.400812222 0.467256702 0.483466027 0.470322681 0.816884728
#>  [55] 0.265663195 0.712062739 0.059984160 0.462630886 0.543449020 0.476628889
#>  [61] 0.500014213 0.218073072 0.368155521 0.089445205 0.130430556 0.603285353
#>  [67] 0.627919346 0.474148021 0.778102426 0.812804433 0.254454257 0.007122939
#>  [73] 0.604120490 0.133386700 0.228913972 0.723120133 0.282153887 0.089693084
#>  [79] 0.404104343 0.314614340 0.396676918 0.520016408 0.308171688 0.525378279
#>  [85] 0.305392207 0.428689720 0.808140639 0.599048136 0.322200044 0.773359411
#>  [91] 0.644667851 0.523766185 0.503334836 0.232742295 0.892002812 0.211838534
#>  [97] 0.884166259 0.788190280 0.370073361 0.134228598 0.236653397 0.517185110
#> [103] 0.391070528 0.274427484 0.114406089 0.352709628 0.197733342 0.043964518
#> [109] 0.287175262 0.805312984 0.279992412 0.668684880 0.042792507 0.418891962
#> [115] 0.684421952 0.509999330 0.284338409 0.211473549 0.189645503 0.145467866
#> [121] 0.497946330 0.154753185 0.289848592 0.396127924 0.899695412 0.182662235
#> [127] 0.439483272 0.502684509 0.136806002 0.456658247 0.899659109 0.666399715
#> [133] 0.455683656 0.264477596 0.017127548 0.275360859 0.047792460 0.718314469
#> [139] 0.817229172 0.067419139 0.482931412 0.282551852 0.046952414 0.046301055
#> [145] 0.049881236 0.176716347 0.069853065 0.720305391 0.586666585 0.094112951
#> [151] 0.271828737 0.725151543 0.533547855 0.052595352 0.385149686 0.327306607
#> [157] 0.369418455 0.337508166 0.602925924 0.341775859 0.647353764 0.135773930
#> [163] 0.048048195 0.466115977 0.331633285 0.464510986 0.707989379 0.267552902
#> [169] 0.630435808 0.640653746 0.255860501 0.374809991 0.388083682 0.793125683
#> [175] 0.205891983 0.135517778 0.341084580 0.619906173 0.548779027 0.310382901
#> [181] 0.098933380 0.686223737 0.343748042 0.077372568 0.344651233 0.330737049
#> [187] 0.724534314 0.529316085 0.542627912 0.306753027 0.549668411 0.351131178
#> [193] 0.526799375 0.138995300 0.066640668 0.741102543 0.687257765 0.076187863
#> [199] 0.160696293 0.117678194 0.627696284 0.798728885 0.329785576 0.400733965
#> [205] 0.310919567 0.303937820 0.183638734 0.189720010 0.734636359 0.212274358
#> [211] 0.495397433 0.460859051 0.801199774 0.244293008 0.155172004 0.836660651
#> [217] 0.329293721 0.210076374 0.091861076 0.065678205 0.250418221 0.486187918
#> [223] 0.801214993 0.491872159 0.327275420 0.412579288 0.126025167 0.230439724
#> [229] 0.759873223 0.096523421 0.807192283 0.308224486 0.434667597 0.211572017
#> [235] 0.269655637 0.089825303 0.401088168 0.550919013 0.109659382 0.200644130
#> [241] 0.205330734 0.282663023 0.571376432 0.116919227 0.411657363 0.531648367
#> [247] 0.424202523 0.086800366 0.127845779 0.649947933 0.281468535 0.161697101
#> [253] 0.246877556 0.458742362 0.740199363 0.263610301 0.696682450 0.493091512
#> [259] 0.428560852 0.059805273 0.239444128 0.147871646 0.475110495 0.392422043
#> [265] 0.753945753 0.862504891 0.426904630 0.036875589 0.334191770 0.399213612
#> [271] 0.465285735 0.621468013 0.746887577 0.078603841 0.645854030 0.228502707
#> [277] 0.554224918 0.469301913 0.576257881 0.199648887 0.046922093 0.534272248
#> [283] 0.590389567 0.377014115 0.532062149 0.541044889 0.483242138 0.878203639
#> [289] 0.166616746 0.481927968 0.838254576 0.680471355 0.568998601 0.257006396
#> [295] 0.864497651 0.278649496 0.598768656 0.085849946 0.397162727 0.408953428

## select a bivariate copula for the simulated data
cop <- BiCopSelect(u1, u2)
summary(cop)

#> Family
#> ------ 
#> No:    16
#> Name:  Survival Joe
#> 
#> Parameter(s)
#> ------------
#> par:  3.81
#> 
#> Dependence measures
#> -------------------
#> Kendall's tau:    0.6 (empirical = 0.57, p value < 0.01)
#> Upper TD:         0 
#> Lower TD:         0.8 
#> 
#> Fit statistics
#> --------------
#> logLik:  165.79 
#> AIC:    -329.59 
#> BIC:    -325.89 
#> 

## and evaluate its CDF
BiCopCDF(u1, u2, cop)

#>   [1] 0.268500182 0.393482100 0.464894789 0.510456272 0.453162024 0.797323487
#>   [7] 0.588601111 0.099482773 0.118512927 0.322486665 0.814900655 0.639423908
#>  [13] 0.599058721 0.475238888 0.184087006 0.940328715 0.650089849 0.020430200
#>  [19] 0.528904262 0.224164034 0.114739901 0.323577868 0.095813832 0.192260638
#>  [25] 0.253748923 0.037373923 0.490382483 0.385460855 0.117787408 0.711378015
#>  [31] 0.384689327 0.280957246 0.672082756 0.768950504 0.651786521 0.672018143
#>  [37] 0.132639547 0.331106236 0.333744533 0.210386952 0.218524522 0.390087202
#>  [43] 0.083562391 0.959755834 0.592526750 0.116640534 0.306823360 0.215342655
#>  [49] 0.452303899 0.393405093 0.459175404 0.481287502 0.466120340 0.813647499
#>  [55] 0.261439988 0.709057371 0.059565885 0.456855357 0.541356648 0.468243272
#>  [61] 0.491857045 0.214263633 0.367249220 0.088220906 0.128309756 0.599365782
#>  [67] 0.621174636 0.466536552 0.774157185 0.811803894 0.249649846 0.007005071
#>  [73] 0.598545572 0.131654073 0.225308946 0.717245500 0.277091702 0.088040456
#>  [79] 0.397385146 0.309124353 0.389332990 0.511528587 0.302444365 0.518453159
#>  [85] 0.299842335 0.422155034 0.804545489 0.590955696 0.316200729 0.768696055
#>  [91] 0.638551588 0.516169624 0.494733169 0.228517466 0.890956219 0.207920208
#>  [97] 0.883587289 0.785326749 0.363739537 0.131843650 0.235281626 0.508567728
#> [103] 0.386962659 0.269377320 0.112533962 0.346416642 0.194429761 0.043262486
#> [109] 0.281730472 0.803591131 0.278380754 0.661849267 0.042005238 0.411355374
#> [115] 0.681672653 0.501380412 0.281113276 0.207511168 0.187186965 0.143467695
#> [121] 0.490187152 0.152096906 0.285201040 0.394537146 0.898689328 0.179480956
#> [127] 0.432015793 0.496324238 0.134272707 0.451968296 0.898589501 0.664577410
#> [133] 0.447749819 0.259512738 0.016816697 0.274800839 0.047021904 0.712741490
#> [139] 0.815875696 0.066286394 0.477457800 0.277572089 0.046088531 0.045501507
#> [145] 0.049059305 0.174202934 0.069094328 0.715817853 0.586037854 0.092815442
#> [151] 0.270848989 0.719645727 0.524842781 0.052324023 0.377958047 0.321111660
#> [157] 0.365998373 0.332792771 0.597147752 0.337683815 0.641655124 0.133597898
#> [163] 0.047588489 0.466104233 0.328637202 0.456727271 0.703639903 0.262539717
#> [169] 0.622767452 0.639739046 0.252030698 0.368368590 0.380839741 0.792276580
#> [175] 0.202299401 0.134794688 0.335626008 0.619286761 0.540158244 0.306364569
#> [181] 0.097274711 0.681772466 0.338758200 0.076805512 0.338154542 0.324523421
#> [187] 0.719092062 0.527588683 0.536167940 0.305369859 0.542888855 0.345900352
#> [193] 0.523276699 0.136491299 0.065504854 0.739952890 0.681850163 0.074856903
#> [199] 0.158411211 0.117409909 0.627132161 0.795021078 0.323532747 0.396392627
#> [205] 0.305673697 0.299958382 0.180254088 0.186509493 0.732255694 0.210431743
#> [211] 0.487512634 0.452879534 0.797313958 0.239692110 0.153348541 0.834670597
#> [217] 0.329071573 0.206381227 0.090183320 0.064623557 0.245902432 0.480025699
#> [223] 0.797301191 0.486220532 0.321714182 0.405018177 0.124537919 0.227980256
#> [229] 0.755207973 0.095071820 0.805946237 0.303203031 0.427218808 0.208086487
#> [235] 0.265929990 0.088650160 0.395098041 0.542244147 0.109549868 0.197394389
#> [241] 0.202540690 0.277935441 0.569151858 0.114998646 0.406585303 0.530848464
#> [247] 0.417431959 0.086630440 0.126422931 0.644136639 0.276155351 0.159563114
#> [253] 0.242750328 0.450544422 0.737566095 0.260132222 0.690973674 0.486846051
#> [259] 0.422447372 0.059011406 0.234950335 0.146394246 0.467648516 0.391205545
#> [265] 0.753383940 0.860474930 0.421434959 0.036314320 0.329816639 0.392135576
#> [271] 0.461163978 0.615344225 0.745823886 0.077665085 0.638724071 0.224563251
#> [277] 0.550017096 0.461276853 0.567681744 0.198301450 0.046534965 0.527854157
#> [283] 0.582504413 0.371944657 0.527726038 0.537312145 0.474778455 0.876594428
#> [289] 0.165391490 0.473484299 0.835710078 0.674601315 0.564915029 0.252633874
#> [295] 0.863535735 0.275419048 0.597632932 0.084917428 0.389853877 0.407451270