Blomqvist's Beta Value of a Bivariate Copula

This function computes the theoretical Blomqvist's beta value of a bivariate copula for given parameter values.

BiCopPar2Beta(family, par, par2 = 0, obj = NULL, check.pars = TRUE)

Arguments

family

integer; single number or vector of size n; defines the bivariate copula family:
0 = independence copula
2 = Student t copula (t-copula)
1 = Gaussian copula
3 = Clayton copula
4 = Gumbel copula
5 = Frank copula
6 = Joe copula
7 = BB1 copula
8 = BB6 copula
9 = BB7 copula
10 = BB8 copula
13 = rotated Clayton copula (180 degrees; survival Clayton'') \cr `14` = rotated Gumbel copula (180 degrees; survival Gumbel'')
16 = rotated Joe copula (180 degrees; survival Joe'') \cr `17` = rotated BB1 copula (180 degrees; survival BB1'')
18 = rotated BB6 copula (180 degrees; survival BB6'')\cr `19` = rotated BB7 copula (180 degrees; survival BB7'')
20 = rotated BB8 copula (180 degrees; ``survival BB8'')
23 = rotated Clayton copula (90 degrees)
`24` = rotated Gumbel copula (90 degrees)
`26` = rotated Joe copula (90 degrees)
`27` = rotated BB1 copula (90 degrees)
`28` = rotated BB6 copula (90 degrees)
`29` = rotated BB7 copula (90 degrees)
`30` = rotated BB8 copula (90 degrees)
`33` = rotated Clayton copula (270 degrees)
`34` = rotated Gumbel copula (270 degrees)
`36` = rotated Joe copula (270 degrees)
`37` = rotated BB1 copula (270 degrees)
`38` = rotated BB6 copula (270 degrees)
`39` = rotated BB7 copula (270 degrees)
`40` = rotated BB8 copula (270 degrees)
`104` = Tawn type 1 copula
`114` = rotated Tawn type 1 copula (180 degrees)
`124` = rotated Tawn type 1 copula (90 degrees)
`134` = rotated Tawn type 1 copula (270 degrees)
`204` = Tawn type 2 copula
`214` = rotated Tawn type 2 copula (180 degrees)
`224` = rotated Tawn type 2 copula (90 degrees)
`234` = rotated Tawn type 2 copula (270 degrees)
Note that the Student's t-copula is not allowed since the CDF of the t-copula is not implemented (see BiCopCDF()).

par

numeric; single number or vector of size n; copula parameter.

par2

numeric; single number or vector of size n; second parameter for the two parameter BB1, BB6, BB7, BB8, Tawn type 1 and type 2 copulas (default: par2 = 0).

obj

BiCop object containing the family and parameter specification.

check.pars

logical; default is TRUE; if FALSE, checks for family/parameter-consistency are omitted (should only be used with care).

Value

Theoretical value of Blomqvist's beta corresponding to the bivariate copula family and parameter(s) par, par2.

Details

If the family and parameter specification is stored in a BiCop() object obj, the alternative version

BiCopPar2Beta(obj)

can be used.

Note

The number n can be chosen arbitrarily, but must agree across arguments.

References

Blomqvist, N. (1950). On a measure of dependence between two random variables. The Annals of Mathematical Statistics, 21(4), 593-600.

Nelsen, R. (2006). An introduction to copulas. Springer

Author

Ulf Schepsmeier

Examples

## Example 1: Gaussian copula
BiCopPar2Beta(family = 1, par = 0.7)
#> [1] 0.4936334
BiCop(1, 0.7)$beta  # alternative
#> [1] 0.4936334

## Example 2: Clayton copula
BiCopPar2Beta(family = 3, par = 2)
#> [1] 0.5118579

## Example 3: different copula families
BiCopPar2Beta(family = c(3,4,6), par = 2:4)
#> [1] 0.5118579 0.6702672 0.6403889