Selection and Maximum Likelihood Estimation of Bivariate Copula Families

This function selects an appropriate bivariate copula family for given bivariate copula data using one of a range of methods. The corresponding parameter estimates are obtained by maximum likelihood estimation.

BiCopSelect(u1, u2, familyset = NA, selectioncrit = "AIC",
  indeptest = FALSE, level = 0.05, weights = NA, rotations = TRUE,
  se = FALSE, presel = TRUE, method = "mle")

Arguments

u1, u2
Data vectors of equal length with values in [0,1].

familyset
Vector of bivariate copula families to select from. The vector has to include at least one bivariate copula family that allows for positive and one that allows for negative dependence. If familyset = NA (default), selection among all possible families is performed. If a vector of negative numbers is provided, selection among all but abs(familyset) families is performed. Coding of bivariate copula families:
0 = independence copula
1 = Gaussian copula
2 = Student t copula (t-copula)
3 = Clayton copula
4 = Gumbel copula
5 = Frank copula
6 = Joe copula
7 = BB1 copula
8 = BB6 copula
9 = BB7 copula
10 = BB8 copula
13 = rotated Clayton copula (180 degrees; ``survival Clayton'')
14 = rotated Gumbel copula (180 degrees; ``survival Gumbel'')
16 = rotated Joe copula (180 degrees; ``survival Joe'')
17 = rotated BB1 copula (180 degrees; ``survival BB1'')
18 = rotated BB6 copula (180 degrees; ``survival BB6'')
19 = rotated BB7 copula (180 degrees; ``survival BB7'')
20 = rotated BB8 copula (180 degrees; ``survival BB8'')
23 = rotated Clayton copula (90 degrees)
24 = rotated Gumbel copula (90 degrees)
26 = rotated Joe copula (90 degrees)
27 = rotated BB1 copula (90 degrees)
28 = rotated BB6 copula (90 degrees)
29 = rotated BB7 copula (90 degrees)
30 = rotated BB8 copula (90 degrees)
33 = rotated Clayton copula (270 degrees)
34 = rotated Gumbel copula (270 degrees)
36 = rotated Joe copula (270 degrees)
37 = rotated BB1 copula (270 degrees)
38 = rotated BB6 copula (270 degrees)
39 = rotated BB7 copula (270 degrees)
40 = rotated BB8 copula (270 degrees)
104 = Tawn type 1 copula
114 = rotated Tawn type 1 copula (180 degrees)
124 = rotated Tawn type 1 copula (90 degrees)
134 = rotated Tawn type 1 copula (270 degrees)
204 = Tawn type 2 copula
214 = rotated Tawn type 2 copula (180 degrees)
224 = rotated Tawn type 2 copula (90 degrees)
234 = rotated Tawn type 2 copula (270 degrees)

selectioncrit
Character indicating the criterion for bivariate copula selection. Possible choices: selectioncrit = "AIC" (default), "BIC", or "logLik".

indeptest
Logical; whether a hypothesis test for the independence of u1 and u2 is performed before bivariate copula selection (default: indeptest = FALSE; see BiCopIndTest). The independence copula is chosen if the null hypothesis of independence cannot be rejected.

level
Numeric; significance level of the independence test (default: level = 0.05).

weights
Numerical; weights for each observation (optional).

rotations
If TRUE, all rotations of the families in familyset are included (or subtracted).

se
Logical; whether standard error(s) of parameter estimates is/are estimated (default: se = FALSE).

presel
Logical; whether to exclude families before fitting based on symmetry properties of the data. Makes the selection about 30% faster (on average), but may yield slightly worse results in few special cases.

method
indicates the estimation method: either maximum likelihood estimation (method = "mle"; default) or inversion of Kendall's tau (method = "itau"). For method = "itau" only one parameter families and the Student t copula can be used (family = 1,2,3,4,5,6,13,14,16,23,24,26,33,34 or 36). For the t-copula, par2 is found by a crude profile likelihood optimization over the interval (2, 10].

Value

An object of class BiCop, augmented with the following entries:

se, se2

standard errors for the parameter estimates (if se = TRUE,

nobs

number of observations,

logLik

log likelihood

AIC

Aikaike's Informaton Criterion,

BIC

Bayesian's Informaton Criterion,

emptau

empirical value of Kendall's tau,

p.value.indeptest

p-value of the independence test.

Details

Copulas can be selected according to the Akaike and Bayesian Information Criteria (AIC and BIC, respectively). First all available copulas are fitted using maximum likelihood estimation. Then the criteria are computed for all available copula families (e.g., if u1 and u2 are negatively dependent, Clayton, Gumbel, Joe, BB1, BB6, BB7 and BB8 and their survival copulas are not considered) and the family with the minimum value is chosen. For observations \(u_{i,j},\ i=1,...,N,\ j=1,2,\) the AIC of a bivariate copula family \(c\) with parameter(s) \(\boldsymbol{\theta}\) is defined as $$AIC := -2 \sum_{i=1}^N \ln[c(u_{i,1},u_{i,2}|\boldsymbol{\theta})] + 2k, $$ where \(k=1\) for one parameter copulas and \(k=2\) for the two parameter t-, BB1, BB6, BB7 and BB8 copulas. Similarly, the BIC is given by $$BIC := -2 \sum_{i=1}^N \ln[c(u_{i,1},u_{i,2}|\boldsymbol{\theta})] + \ln(N)k. $$ Evidently, if the BIC is chosen, the penalty for two parameter families is stronger than when using the AIC.

Additionally a test for independence can be performed beforehand.

Note

For a comprehensive summary of the fitted model, use summary(object); to see all its contents, use str(object).

The parameters of the Student t and BB copulas are restricted (see defaults in BiCopEst to avoid being to close to their limiting cases.

References

Akaike, H. (1973). Information theory and an extension of the maximum likelihood principle. In B. N. Petrov and F. Csaki (Eds.), Proceedings of the Second International Symposium on Information Theory Budapest, Akademiai Kiado, pp. 267-281.

Brechmann, E. C. (2010). Truncated and simplified regular vines and their applications. Diploma thesis, Technische Universitaet Muenchen.
http://mediatum.ub.tum.de/?id=1079285.

Manner, H. (2007). Estimation and model selection of copulas with an application to exchange rates. METEOR research memorandum 07/056, Maastricht University.

Schwarz, G. E. (1978). Estimating the dimension of a model. Annals of Statistics 6 (2), 461-464.

See also

Examples

## Example 1: Gaussian copula with large dependence parameter
par <- 0.7
fam <- 1
dat1 <- BiCopSim(500, fam, par)
# select the bivariate copula family and estimate the parameter(s)
cop1 <- BiCopSelect(dat1[, 1], dat1[, 2], familyset = 1:10,
                    indeptest = FALSE, level = 0.05)
cop1  # short overview

#> Bivariate copula: Gaussian (par = 0.68, tau = 0.47) 

summary(cop1)  # comprehensive overview

#> Family
#> ------ 
#> No:    1
#> Name:  Gaussian
#> 
#> Parameter(s)
#> ------------
#> par:  0.68
#> 
#> Dependence measures
#> -------------------
#> Kendall's tau:    0.47 (empirical = 0.46, p value < 0.01)
#> Upper TD:         0 
#> Lower TD:         0 
#> 
#> Fit statistics
#> --------------
#> logLik:  140.06 
#> AIC:    -278.11 
#> BIC:    -273.9 
#> 

str(cop1)  # see all contents of the object

#> List of 15
#>  $ family           : num 1
#>  $ par              : num 0.675
#>  $ par2             : num 0
#>  $ npars            : num 1
#>  $ familyname       : chr "Gaussian"
#>  $ tau              : num 0.472
#>  $ beta             : num 0.472
#>  $ taildep          :List of 2
#>   ..$ lower: num 0
#>   ..$ upper: num 0
#>  $ call             : language BiCopSelect(u1 = dat1[, 1], u2 = dat1[, 2], familyset = 1:10, indeptest = FALSE,      level = 0.05)
#>  $ nobs             : int 500
#>  $ logLik           : num 140
#>  $ AIC              : num -278
#>  $ BIC              : num -274
#>  $ emptau           : num 0.461
#>  $ p.value.indeptest: num 0
#>  - attr(*, "class")= chr "BiCop"

## Example 2: Gaussian copula with small dependence parameter
par <- 0.01
fam <- 1
dat2 <- BiCopSim(500, fam, par)
# select the bivariate copula family and estimate the parameter(s)
cop2 <- BiCopSelect(dat2[, 1], dat2[, 2], familyset = 0:10,
                    indeptest = TRUE, level = 0.05)
summary(cop2)

#> Family
#> ------ 
#> No:    5
#> Name:  Frank
#> 
#> Parameter(s)
#> ------------
#> par:  -0.58
#> 
#> Dependence measures
#> -------------------
#> Kendall's tau:    -0.06 (empirical = -0.07, p value = 0.02)
#> Upper TD:         0 
#> Lower TD:         0 
#> 
#> Fit statistics
#> --------------
#> logLik:  2.51 
#> AIC:    -3.02 
#> BIC:    1.19 
#> 

## Example 3: empirical data
data(daxreturns)
cop3 <- BiCopSelect(daxreturns[, 1], daxreturns[, 4], familyset = 0:10)
summary(cop3)

#> Family
#> ------ 
#> No:    14
#> Name:  Survival Gumbel
#> 
#> Parameter(s)
#> ------------
#> par:  1.61
#> 
#> Dependence measures
#> -------------------
#> Kendall's tau:    0.38 (empirical = 0.4, p value < 0.01)
#> Upper TD:         0 
#> Lower TD:         0.46 
#> 
#> Fit statistics
#> --------------
#> logLik:  241.84 
#> AIC:    -481.68 
#> BIC:    -476.62 
#>