Visualisation of R-Vine Tree Structure

Function is deprecated since VineCopula 2.0. Use plot.RVineMatrix instead.

RVineTreePlot(x, tree = "ALL", type = 0, edge.labels = NULL,
  legend.pos = "bottomleft", interactive = FALSE, ...)

Arguments

x
RVineMatrix object.

tree
"ALL" or integer vector; specifies which trees are plotted.

type
integer; specifies how to make use of variable names: 0 = variable names are ignored, 1 = variable names are used to annotate vertices, 2 = uses numbers in plot and adds a legend for variable names.

edge.labels
character; either a vector of edge labels or one of the following: "family" = pair-copula family abbreviation (see BiCopName), "par" = pair-copula parameters, "tau" = pair-copula Kendall's tau (by conversion of parameters) "family-par" = pair-copula family and parameters "family-tau" = pair-copula family and Kendall's tau.

legend.pos
the x argument for legend.

interactive
logical; if TRUE, the user is asked to adjust the positioning of vertices with his mouse.

…
Arguments passed to plot.network.

See also

plot.RVineMatrix